Role authorization

+++ src/Cde.Api/Controllers/UserController.cs
@@ -14,6 +14,8 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using AutoMapper;
 using Cde.Models.DTOs;
+using System.Security.Claims;
+using Cde.Api.Constants;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -26,22 +28,27 @@ namespace Cde.Controllers
 	public class UserController : ControllerBase
 	{
 		private readonly DatabaseService<UserModel> userService;
+		private readonly DatabaseService<RoleModel> roleService;
+		private readonly IMapper _mapper;
 
-		public UserController(ApplicationDbContext context) {
-			userService = new DatabaseService<UserModel>(context);
+		public UserController(ApplicationDbContext context, IMapper mapper) {
+			userService = new UserService(context);
+			roleService = new DatabaseService<RoleModel>(context);
+			_mapper = mapper;
 		}
 
 		// GET: api/<UserController>
 		[HttpGet]
-		public ActionResult<List<UserModel>> GetAllUsers() {
-			return Ok(userService.GetAll().ToList());
+		[Authorize(Roles.Admin)]
+		public ActionResult<List<UserDTO>> GetAllUsers() {
+			return Ok(_mapper.Map<List<UserDTO>>(userService.GetAll().ToList()));
 		}
 
 		// GET api/<UserController>/5
 		[HttpGet("{id}")]
-		public ActionResult<UserModel> GetUser(int id) {
+		public ActionResult<UserDTO> GetUser(int id) {
 			try {
-				return Ok(userService.Get(u => u.Id == id).First());
+				return Ok(_mapper.Map<UserDTO>(userService.Get(u => u.Id == id).First()));
 			} catch (Exception e) {
 				if (e is ArgumentNullException || e is InvalidOperationException) {
 					return NotFound(new { error = "User not found" });
@@ -52,41 +59,52 @@ namespace Cde.Controllers
 
 		// POST api/<UserController>
 		[HttpPost]
+		[Authorize(Roles.Admin)]
 		public ActionResult<UserModel> Post([FromBody] UserFormDTO userForm) {
-			if (null != userService.Get(u => u.Email == userForm.Email).FirstOrDefault()) {
+			if (null != userService.Get(u => u.Name == userForm.Name).FirstOrDefault()) {
 				return Conflict(new { error = "User alredy exist!" });
+			}
+			var role = roleService.Get(r => string.Equals(r.Name, userForm.Role, StringComparison.OrdinalIgnoreCase)).FirstOrDefault();
+			if (null == role) {
+				return BadRequest(new { error = "Role does not exist" });
 			}
 			UserModel user = new UserModel() {
 				Name = userForm.Name,
-				Email = userForm.Email,
+				RoleId = role.Id,
 				CreatedAt = DateTime.UtcNow,
-				Salt = PasswordManager.GenerateSalt(userForm.Email)
+				Salt = PasswordManager.GenerateSalt(userForm.Name)
 			};
 			user.Passhash = PasswordManager.GeneratePasshash(user.Salt, userForm.Password);
 			return Created("", userService.Create(user));
 		}
 
 		// PUT api/<UserController>/5
-		[HttpPut("{id}")]
-		public ActionResult<UserModel> Put(int id, [FromBody] UserModel user) {
+		[HttpPut("role/{id}")]
+		[Authorize(Roles.Admin)]
+		public ActionResult<UserModel> PutUpdateRole(int id, [FromBody] string roleName) {
 			var updatedUser = userService.Get(u => u.Id == id).FirstOrDefault();
 			if (null == updatedUser) {
 				return NotFound(new { error = "User not found" });
 			}
-			updatedUser.Id = id;
-			updatedUser.Name = user.Name;
-			updatedUser.Email = user.Email;
+			var role = roleService.Get(r => string.Equals(r.Name, roleName, StringComparison.OrdinalIgnoreCase)).FirstOrDefault();
+			if (null == role) {
+				return BadRequest(new { error = "Role does not exist" });
+			}
+			updatedUser.RoleId = role.Id;
 			return Ok(userService.Update(updatedUser));
 		}
 
 		[HttpPut("password/{id}")]
-		public ActionResult<UserModel> Put(int id, [FromBody] UserFormDTO userForm) {
+		public ActionResult<UserModel> PutUpdatePass(int id, [FromBody] string password) {
 			var updatedUser = userService.Get(u => u.Id == id).FirstOrDefault();
 			if (null == updatedUser) {
 				return NotFound(new { error = "User not found" });
 			}
-			updatedUser.Salt = PasswordManager.GenerateSalt(updatedUser.Email);
-			updatedUser.Passhash = PasswordManager.GeneratePasshash(updatedUser.Salt, userForm.Password);
+			if (updatedUser.Name != User.FindFirst(ClaimTypes.Name).Value) {
+				return BadRequest(new { error = "Cannot update password from another user" });
+			}
+			updatedUser.Salt = PasswordManager.GenerateSalt(updatedUser.Name);
+			updatedUser.Passhash = PasswordManager.GeneratePasshash(updatedUser.Salt, password);
 			return Ok(userService.Update(updatedUser));
 		}
 
@@ -97,6 +115,9 @@ namespace Cde.Controllers
 			if (null == user) {
 				return NotFound(new { error = "User not found" });
 			}
+			if (user.Name != User.FindFirst(ClaimTypes.Name).Value || !UserHelper.IsAdim(User)) {
+				return BadRequest(new { error = "Cannot delete user" });
+			}
 			userService.Delete(user);
 			return Ok();
 		}
@@ -104,7 +125,7 @@ namespace Cde.Controllers
 		[AllowAnonymous]
 		[HttpPost("authenticate")]
 		public ActionResult<AuthenticateResponseDTO> Authenticate([FromBody] AuthenticateRequestDTO model) {
-			var dbUser = userService.Get(u => u.Email == model.Email).FirstOrDefault();
+			var dbUser = userService.Get(u => u.Name == model.Name).FirstOrDefault();
 			if (null == dbUser) {
 				return BadRequest(new { error = "User does not exist" });
 			} else if (PasswordManager.GeneratePasshash(dbUser.Salt, model.Password) != dbUser.Passhash) {
